Fixed program crashes caused by the deprecation of matplotlib's resize_event() and renamed program file
The program has been broken for a while because the command fig.canvas.resize_event(), which the program relied on to update figure titles is deprecated. I optimized the program to work with different non-deprecated functions. I also renamed it from display_all_end_cities.py to end_city_autopilot.py to reflect the additional progression and level of completeness of the program. It is now fully operational!